requirements.txt

---

### `requirements.txt`

```plaintext
openai
python-dotenv
requests